Fix cloasing modal after sending data

+++ components/files/modals/uploadFiles/ModalFooter.tsx
@@ -3,7 +3,7 @@ import { DialogClose, DialogFooter } from "@/components/ui/dialog";
 import { deleteAllFilesFromIndexedDB } from "@/lib/indexedDB";
 import { deleteFile } from "@/services/deleteFile";
 import { useUploadStore } from "@/store/uploadFileStore";
-import React, { ReactElement } from "react";
+import React, { ReactElement, useRef } from "react";
 import toast from "react-hot-toast";
 import { useMutation } from "@tanstack/react-query";
 import { Loader2 } from "lucide-react";
@@ -17,8 +17,10 @@ type CleanFileType = {
 };
 
 export default function ModalFooter(): ReactElement {
+  // Create ref for hidden btn, for using closing modal after sending data
+  const closeRef = useRef<HTMLButtonElement>(null);
   const { files, clearFiles } = useUploadStore();
-  toast.error("خطا! اپلود فایل ها را کامل کنید");
+
   // Send data mutation
   const mutation = useMutation<void, Error, CleanFileType[]>({
     mutationFn: async (files: CleanFileType[]) => {
@@ -34,12 +36,15 @@ export default function ModalFooter(): ReactElement {
       clearFiles();
       deleteAllFilesFromIndexedDB();
       toast.success("فایل‌ها با موفقیت ذخیره شدند");
+      // Click on hidden closing btn after sending data
+      closeRef.current?.click();
     },
     onError: () => toast.error("ارسال فایل‌ها ناموفق بود"),
   });
 
   // Clear all uploaded file after click on "انصراف" btn
   const handleClearFiles = () => {
+    if (files.length === 0) return;
     files.forEach(async (f) => {
       await deleteFile(f.status, f.id, f.url);
     });
@@ -74,30 +79,36 @@ export default function ModalFooter(): ReactElement {
     mutation.mutate(cleanFiles);
   };
   return (
-    <DialogFooter className="md:justify-between">
+    <>
       <DialogClose asChild>
-        <Button
-          onClick={handleClearFiles}
-          variant="outline"
-          className="md:w-[50%] border-0 bg-gray-200 items-center dark:bg-[var(--tertiary-dark)] dark:text-white"
-        >
-          انصراف
-        </Button>
+        <button ref={closeRef} className="hidden"></button>
       </DialogClose>
-      <Button
-        onClick={handleSendData}
-        variant="outline"
-        className="md:w-[50%] border-0 bg-[var(--secondary)] items-center text-white"
-      >
-        {mutation.isPending ? (
-          <>
-            <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-            در حال ارسال
-          </>
-        ) : (
-          "تایید"
-        )}
-      </Button>
-    </DialogFooter>
+
+      <DialogFooter className="md:justify-between">
+        <DialogClose asChild>
+          <Button
+            onClick={handleClearFiles}
+            variant="outline"
+            className="md:w-[50%] border-0 bg-gray-200 items-center dark:bg-[var(--tertiary-dark)] dark:text-white"
+          >
+            انصراف
+          </Button>
+        </DialogClose>
+        <Button
+          onClick={handleSendData}
+          variant="outline"
+          className="md:w-[50%] border-0 bg-[var(--secondary)] items-center text-white"
+        >
+          {mutation.isPending ? (
+            <>
+              <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+              در حال ارسال
+            </>
+          ) : (
+            "تایید"
+          )}
+        </Button>
+      </DialogFooter>
+    </>
   );
 }
